Add feature to test all kata setup.sh scripts
./utils/test/test_setup.sh can be used to test a single kata
Run it from the kata folder as:
../utils/test/test_setup.sh

To test all katas, run ./test-all.sh from the root folder.

This commit also includes a few bug-fixes that were caught by the test.
In reverted-merge there is a merge operation in the setup that causes a
conflict on purpose, and fixes it.
I "wrapped" that in an `|| true` to make it not fail the test script.